fix: rename frozen table
